ui,e2e: move cypress helper function into util file after upgrade to 13
The previous upgrade fixed the race condition but introduced an issue with the file
organization that is new to cypress v13.

This change moves the shared helper function out into a new file to resolve the test failure.
This change also removes the column header interaction from the insights page to remove a
flakey failure.

Fixes #164240and fixes #164239
Epic: None
Release Note: None
