+++ pkg/ui/workspaces/e2e-tests/cypress/e2e/health-check/authenticated.cy.ts
@@ -4,7 +4,7 @@
 // included in the /LICENSE file.
 
 import { SQLPrivilege } from "../../support/types";
-import { isTextGreaterThanZero } from "../../support/e2e";
+import { isTextGreaterThanZero } from "../../support/utils";
 
 describe("health check: authenticated user", () => {
   it("serves a DB Console overview page", () => {
